+++ CHANGELOG.md
@@ -50,6 +50,7 @@
 - Other change in `tramtracker.types`:
   - In `TramDeparture`, added new API data fields: `location`, `occupancy_level` and `AVM_timestamp`
 - The first parameter (`cls`) of class methods is now of type `type[Self]` instead of the erroneous `Self`
+- Increased API rate limits of `ptv_timetable` and `tramtracker` to 20 calls per 60 seconds
 - Fixed bugs in `ptv_timetable.TimetableAPI` and `ptv_timetable.asyncapi.AsyncTimetableAPI`:
   - `geopath_date` value in `get_route()` was not converted to string before being sent to the API
   - A `None` argument check for the `date` parameter was missing in `get_run()`